Added SHA1 SPKI hash.
