+++ VCSJones.FiddlerCert.sln
@@ -8,9 +8,6 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VCSJones.FiddlerCert.UnitTests", "VCSJones.FiddlerCert.UnitTests\VCSJones.FiddlerCert.UnitTests.csproj", "{7EBF08C8-303F-4784-85AB-C3C00CFF5A5F}"
 EndProject
 Global
-	GlobalSection(Performance) = preSolution
-		HasPerformanceSessions = true
-	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
