+++ VCSJones.FiddlerCert/CallerMemberNameAttribute.cs
@@ -1,0 +1,9 @@
+﻿ // ReSharper disable once CheckNamespace
+namespace System.Runtime.CompilerServices
+{
+    [AttributeUsage(AttributeTargets.Parameter, Inherited = false)]
+    public class CallerMemberNameAttribute : Attribute
+    {
+        
+    }
+}
